+++ src/modules/ListFood.js
@@ -1,7 +1,7 @@
 import getFood from './getFood.js';
-import { addLikes, popUpComment } from './addLikes.js';
 import counter from './counter.js';
-import { homePage } from './fetchData.js'
+import { addLike, homePage, fetchComments } from './fetchData.js';
+import loadPopUpComment from './comments.js';
 
 const getOject = (arr) => {
   const ob = {};
@@ -11,8 +11,31 @@ const getOject = (arr) => {
   return ob;
 };
 
+const addLikes = () => {
+  const likes = document.querySelectorAll('.like');
+  likes.forEach((like) => {
+    like.addEventListener('click', (e) => {
+      const id = e.target.id.slice(4);
+      addLike({ item_id: id });
+      listFood();
+    });
+  });
+};
+
+const popUpComment = () => {
+  const comments = document.querySelectorAll('.comments');
+  comments.forEach((comment) => {
+    comment.addEventListener('click', (e) => {
+      const id = e.target.id.slice(7);
+      fetchComments(id).then((data) => {
+        loadPopUpComment(data);
+      });
+    });
+  });
+};
+
 const listFood = () => {
-  homePage().then(({foodlist, likesArr}) => {
+  homePage().then(({ foodlist, likesArr }) => {
     const list = document.getElementById('food-list');
     list.innerHTML = '';
     const objectkey = getOject(likesArr);
@@ -23,14 +46,14 @@ const listFood = () => {
       newLi.setAttribute('id', `food${food.idCategory}`);
       const likes = objectkey[foodId];
       if (foodId in objectkey) {
-        newLi.innerHTML = getFood(food, likes)
-      }else {
+        newLi.innerHTML = getFood(food, likes);
+      } else {
         newLi.innerHTML = getFood(food);
       }
       list.appendChild(newLi);
     });
     addLikes();
     popUpComment();
-  })
+  });
 };
 export default listFood;
